pyproject.toml
[project]
name = "culinary-portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 120

lint.select = [
    "ALL",
]

lint.ignore = [
    # modules
    "ANN",

    # specific rules
    "D100",
    "D101",
    "D102",
    "D104",
    "D105",
    "D106",
    "D203",
    "D212",
    "F401",
    "RUF012",
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.per-file-ignores]
"cooking/urls.py" = ["F405", "F403"]
